+++ CHANGELOG.md
@@ -53,7 +53,7 @@
 
 ##### Implement get satoshi per byte functionality
 
-- Added getSatPerByte() to fetch the satoshi per byte rate
+- Standardize getFees() to provide transaction fee in sathoshi and transaction size
 - Added HD PATH for testnet
 - Updated tests to add assertions
 - Added constants for base URLs